requirements.txt

streamlit
openai
textblob
pandas
python-dotenv